+++ client/driver/executor/resource_container_default.go
@@ -1,0 +1,18 @@
+// +build darwin dragonfly freebsd netbsd openbsd solaris windows
+
+package executor
+
+import (
+	dstructs "github.com/hashicorp/nomad/client/driver/structs"
+)
+
+type resourceContainer struct {
+}
+
+func (rc *resourceContainer) cleanup() error {
+	return nil
+}
+
+func (rc *resourceContainer) getIsolationConfig() *dstructs.IsolationConfig {
+	return nil
+}
