+++ client/driver/structs/structs_default.go
@@ -1,0 +1,6 @@
+// +build darwin dragonfly freebsd netbsd openbsd solaris windows
+
+package structs
+
+type IsolationConfig struct {
+}
